♻️  F15: FORM > Create Inputs to extract all TextInput and SelectInput
