+++ src/components/pages/Order/Main/MainRightSide/AdminPanel/Form.jsx
@@ -2,52 +2,31 @@
 /* eslint-disable react/prop-types */
 import React from "react";
 import { styled } from "styled-components";
-import TextInput from "../../../../../reusable-ui/TextInput";
 import ImagePreview from "./ImagePreview";
-import { getInputTextsConfig, getSelectInputConfig } from "./inputConfig";
-import SelectInput from "../../../../../reusable-ui/SelectInput";
+import { Inputs } from "./Inputs";
 
 const Form = React.forwardRef(
   ({ product, onSubmit, onChange, onFocus, onBlur, children }, ref) => {
     // state
 
-    const inputTexts = getInputTextsConfig(product);
-    const inputSelects = getSelectInputConfig(product);
     // affichage
     return (
       <FormStyled onSubmit={onSubmit}>
         <ImagePreview imageSource={product.imageSource} title={product.title} />
-        <div className="input-fields">
-          {inputTexts.map((input) => (
-            <TextInput
-              key={input.id}
-              {...input}
-              onChange={onChange}
-              version="minimalist"
-              onFocus={onFocus}
-              onBlur={onBlur}
-              ref={ref && input.name === "title" ? ref : null}
-            />
-          ))}
-
-          {inputSelects.map((inputSelect) => (
-            <SelectInput
-              {...inputSelect}
-              key={inputSelect.id}
-              onChange={onChange}
-              onFocus={onFocus}
-              onBlur={onBlur}
-            />
-          ))}
-        </div>
-        <div className="submit">{children}</div>
+        <Inputs
+          product={product}
+          onChange={onChange}
+          onFocus={onFocus}
+          onBlur={onBlur}
+          ref={ref}
+        />
+        <div className="form-footer">{children}</div>
       </FormStyled>
     );
   }
 );
 export default Form;
 const FormStyled = styled.form`
-  /* border: 2px solid black; */
   display: grid;
   grid-template-columns: 1fr 3fr;
   grid-template-rows: repeat(4, 1fr);
@@ -56,28 +35,7 @@ const FormStyled = styled.form`
   grid-column-gap: 20px;
   grid-row-gap: 8px;
 
-  .input-fields {
-    /* background: blue; */
-    grid-area: 1 / 2 / -2 / 3;
-    display: grid;
-    grid-template-rows: repeat(3, 1fr);
-    grid-template-columns: repeat(3, 1fr);
-    grid-row-gap: 8px;
-    grid-column-gap: 8px;
-
-    .title {
-      grid-area: 1 / 1 / 2 / 4;
-    }
-    .image-source {
-      grid-area: 2 / 1 / 3 / 4;
-    }
-    .price {
-      grid-area: 3 / 1 / 4 / 2;
-    }
-    
-  }
-
-  .submit {
+  .form-footer {
     /* background: green; */
     grid-area: 4 / -2 / -1 / -1;
     display: flex;
